+++ packages/components/modules/comments/web/CommentsList/EmptyState/index.tsx
@@ -1,0 +1,3 @@
+const EmptyState = () => null
+
+export default EmptyState